refactor: replace userSevice
